ai/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
python-multipart==0.0.9
openai==1.46.0
python-dotenv==1.0.1
xgboost==3.0.5
lightgbm==4.6.0